backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-celery-beat==2.6.0
django-storages==1.14.2
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
weasyprint==61.2
openai==1.35.0
pillow==10.3.0
twilio==9.0.0
python-decouple==3.8
gunicorn==22.0.0
